+++ src/app/authentication.js
@@ -110,6 +110,38 @@ class Authentication {
       }
     })
   }
+
+  async refreshToken() {
+    if (hybrid.isNativeApp()) {
+      const updatedToken = Promise.race([
+        new Promise((resolve) => {
+          hybrid.on('authenticated', (token) => resolve(token))
+        }),
+        new Promise((_, reject) => {
+          setTimeout(() => reject(new Error('Timeout: no token received')), 2000)
+        }),
+      ])
+
+      hybrid.call('refreshExpiredToken')
+
+      return (await updatedToken).radioToken
+    } else {
+      try {
+        const response = await fetch('/login/refresh', {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+          },
+        })
+        const token = (await response.json()).radioToken
+        localStorage.setItem(RADIO_TOKEN_LOCAL_STORAGE_KEY, token)
+        return token
+      } catch (e) {
+        localStorage.removeItem(RADIO_TOKEN_LOCAL_STORAGE_KEY)
+        throw e
+      }
+    }
+  }
 }
 
 export default new Authentication()
